+++ app/tests/components_tests/test_views.py
@@ -1,0 +1,30 @@
+import pytest
+from django.conf import settings
+
+from tests.factories import UserFactory
+from tests.utils import get_view_for_user
+
+
+@pytest.mark.django_db
+class TestComponentInterfaceListView:
+    def test_login_required(self, client):
+        def _get_view(user, context):
+            return get_view_for_user(
+                client=client,
+                viewname=f"components:component-interface-list-{context}",
+                user=user,
+            )
+
+        for context in (
+            "algorithms",
+            "archives",
+            "reader-studies",
+            "input",
+            "output",
+        ):
+            response = _get_view(user=None, context=context)
+            assert response.status_code == 302
+            assert settings.LOGIN_URL in response.url
+
+            response = _get_view(user=UserFactory(), context=context)
+            assert response.status_code == 200
